pyproject.toml
[project]
name = "aimbc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.18",
    "ipykernel>=7.1.0",
    "neo4j>=6.0.3",
    "pdfplumber>=0.11.8",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "deepagents>=0.3.1",
    "tavily-python>=0.7.17",
    "langchain>=1.2.0",
    "anthropic>=0.75.0",
    "langchain-openai>=0.3.0",
    "markdownify>=0.11.0",
    "langgraph-cli[inmem]>=0.4.11",
    "mcp>=1.25.0",
    "langchain-mcp-adapters>=0.2.1",
    "h2>=4.3.0",
]
